PP-5978: Move the PerformanceDao to the 'uk.gov.pay.ledger.report' package
